Refactor Activity Engine
